LayoutTests/inspector/elements/shadow-dom-modify-chardata-expected.txt

Tests that elements panel updates shadow dom tree structure upon typing.



Running: testDumpInitial
========= Original ========
- <div id="container">
    - <input type="text" id="input1">
        - #document-fragment
          <div></div>
      </input>
  </div>

Running: testAppend
======== Type text =========
- <div id="container">
    - <input type="text" id="input1">
        - #document-fragment
        + <div>Bar</div>
      </input>
  </div>
